Installer/*: Check VBOX_WITH_FE_BUGREPORT before assuming VBoxBugReport is there. bugref:10030
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@98676 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Installer/linux/install.sh
@@ -333,7 +333,9 @@ if [ "$ACTION" = "install" ]; then
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxVRDP
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxHeadless
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBalloonCtrl
-    ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBugReport
+    if [ -f $INSTALLATION_DIR/VBoxBugReport ]; then
+        ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBugReport
+    fi
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxAutostart
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/vboxwebsrv
     ln -sf $INSTALLATION_DIR/vbox-img /usr/bin/vbox-img
@@ -368,7 +370,9 @@ if [ "$ACTION" = "install" ]; then
     ln -sf VBoxManage /usr/bin/vboxmanage > /dev/null 2>&1
     ln -sf VBoxSDL /usr/bin/vboxsdl > /dev/null 2>&1
     ln -sf VBoxHeadless /usr/bin/vboxheadless > /dev/null 2>&1
-    ln -sf VBoxBugReport /usr/bin/vboxbugreport > /dev/null 2>&1
+    if [ -f $INSTALLATION_DIR/VBoxBugReport ]; then
+        ln -sf VBoxBugReport /usr/bin/vboxbugreport > /dev/null 2>&1
+    fi
     if [ -f $INSTALLATION_DIR/VBoxDTrace ]; then
         ln -sf VBoxDTrace /usr/bin/vboxdtrace > /dev/null 2>&1
     fi
